// src/pages/assets/types.ts
export type ProfileType =
    | "kprofile"
    | "kplayer"
    | "kpartner";

export type ActiveComponent =
    | "dashboard"
    | "competence"
    | "bookmark"
    | "contact"
    | "missions"
    | "settings"
    | "company"
    | "search"
    | "profile"
    | "contacts"
    | "fileText1"
    | "target"
    | "user"
    | null;